+++ src/core/EventBus.ts
@@ -1,12 +1,14 @@
-import { EventDispatcher } from './EventDispatcher';
-import { EventPayload } from './EventPayload';
-import { LocalEventDriver } from './drivers/LocalEventDriver';
+import debug from 'debug';
+
 import {
-	EventBusOptions,
 	EventDispatcherResponse,
 	EventDriverInterface,
+	EventBusOptions,
 	EventHandler,
 } from './types';
+import { LocalEventDriver } from './drivers/LocalEventDriver';
+import { EventDispatcher } from './EventDispatcher';
+import { EventPayload } from './EventPayload';
 
 /**
  * @file Event bus to subscribe, unsubscribe and publish events.
@@ -37,16 +39,29 @@ export class EventBus {
 	private _drivers: Map<string, EventDriverInterface>;
 
 	/**
+	 * Ongoing promises.
+	 *
+	 * @type {Set<Promise<EventDispatcherResponse>>}
+	 * @private
+	 * @since 3.0.0
+	 * @memberof EventBus
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	private _ongoing: Set<Promise<EventDispatcherResponse>>;
+
+	/**
 	 * Construct with default driver.
 	 *
 	 * @constructor
 	 * @private
 	 * @since 1.0.0
+	 * @since 3.0.0 Added promise tracking.
 	 * @memberof EventBus
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	private constructor() {
 		this._drivers = new Map();
+		this._ongoing = new Set();
 		this.register(new LocalEventDriver());
 	}
 
@@ -69,6 +84,69 @@ export class EventBus {
 	}
 
 	/**
+	 * Get the number of ongoing promises.
+	 *
+	 * @returns {number}
+	 * @public
+	 * @since 3.0.0
+	 * @memberof EventBus
+	 */
+	public get ongoing(): number {
+		return this._ongoing.size;
+	}
+
+	/**
+	 * The cleanup method will:
+	 *
+	 * - Wait for all pending promises triggered by send method.
+	 * - Clear the pool of pending promises.
+	 *
+	 * Useful for:
+	 *
+	 * - Flush pool of pending promises.
+	 * - Ensure all promises are settled.
+	 * - Gracefully shutdown your application.
+	 *
+	 * @returns {Promise<void>}
+	 * @public
+	 * @since 3.0.0
+	 * @memberof EventBus
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	public async cleanup(): Promise<void> {
+		debug('eventbus:cleanup')('cleaning up');
+		await Promise.allSettled(this._ongoing);
+		this._ongoing = new Set();
+		debug('eventbus:cleanup')('cleaned up');
+	}
+
+	/**
+	 * Publish an event to event bus, where Event is an event payload object.
+	 *
+	 * @param {Event} event Event payload object.
+	 * @param {EventBusOptions} options With driver name.
+	 * @returns {Promise<EventDispatcherResponse>}
+	 * @public
+	 * @since 1.0.0
+	 * @memberof EventBus
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	public async publish<Event extends EventPayload>(
+		event: Event,
+		options?: EventBusOptions
+	): Promise<EventDispatcherResponse> {
+		const driver = this.driver(options);
+		const dispatcher = driver.get(event.name);
+
+		if (dispatcher === undefined) {
+			return [];
+		}
+
+		debug('eventbus:publish')(`published ${event.name}`);
+		return dispatcher.dispatch<Event>(event);
+	}
+
+	/**
 	 * Register a new driver for event bus.
 	 *
 	 * @param {EventDriverInterface} driver Event driver object.
@@ -83,28 +161,85 @@ export class EventBus {
 	}
 
 	/**
-	 * Publish an event to event bus, where Event is an event payload object.
+	 * Send an event to event bus, where Event is an event payload object.
+	 * It will not wait for the event to be processed. But:
 	 *
+	 * - When error callback is set on driver, it will be called when something goes wrong.
+	 * - When wait_onclean is true, the event will be added to the pool.
+	 *   a) Later, you may use cleanup method to wait for all pending promises on pool.
+	 *   b) Will be useful for wait_onclean events that must be processed and waited to finish.
+	 *   c) In a graceful shutdown, for example.
+	 *
+	 * Tip: You may want to use wait_onclean for events that should be completed on a
+	 *      graceful shutdown or interruption, for example. It will heavily increase
+	 *      memory usage if you have a lot of events to wait.
+	 *
+	 * @important Be careful when using wait_onclean, it will hold promises on memory till they are settled.
+	 * @experimental This feature is experimental and may change in the future.
 	 * @param {Event} event Event payload object.
+	 * @param {boolean} wait_onclean If true, the event will be sent as wait_onclean.
 	 * @param {EventBusOptions} options With driver name.
-	 * @returns {boolean}
+	 * @returns {void}
 	 * @public
 	 * @since 1.0.0
+	 * @since 3.0.0 Added promise tracking.
 	 * @memberof EventBus
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
-	public async publish<Event extends EventPayload>(
+	public send<Event extends EventPayload>(
 		event: Event,
+		wait_onclean: boolean = false,
 		options?: EventBusOptions
-	): Promise<EventDispatcherResponse> {
+	): void {
 		const driver = this.driver(options);
 		const dispatcher = driver.get(event.name);
 
 		if (dispatcher === undefined) {
-			return undefined;
+			return;
 		}
 
-		return dispatcher.dispatch<Event>(event);
+		if (wait_onclean === false) {
+			dispatcher
+				.dispatch<Event>(event)
+				.then(e => {
+					e.forEach(r => {
+						if (r.status === 'fulfilled') {
+							debug('eventbus:publish')(`sent.resolved ${event.name}`);
+						} else {
+							driver.error(r.reason);
+						}
+					});
+				})
+				.catch(driver.error);
+
+			return;
+		}
+
+		const promise = dispatcher.dispatch<Event>(event);
+
+		this._ongoing.add(promise);
+		debug('eventbus:publish')(`sent ${event.name}; pool: ${this._ongoing.size}`);
+
+		const handleFinally = () => {
+			this._ongoing.delete(promise);
+
+			debug('eventbus:publish')(
+				`send.removed ${event.name}; pool: ${this._ongoing.size}`
+			);
+		};
+
+		promise
+			.then(e => {
+				e.forEach(r => {
+					if (r.status === 'fulfilled') {
+						debug('eventbus:publish')(`sent.resolved ${event.name}`);
+					} else {
+						driver.error(r.reason);
+					}
+				});
+			})
+			.catch(driver.error)
+			.finally(handleFinally);
 	}
 
 	/**
@@ -185,6 +320,7 @@ export class EventBus {
 	}
 
 	/**
+	 * Get driver instance.
 	 *
 	 * @param {EventBusOptions} options
 	 * @returns {EventDriverInterface}