+++ src/core/EventDispatcher.ts
@@ -1,6 +1,6 @@
-import { EventPayload } from './EventPayload';
-
 import type { EventDispatcherResponse, EventHandler } from './types';
+
+import { EventPayload } from './EventPayload';
 
 /**
  * @file Event dispatcher, where you can register handlers to an unique event and dispatch to them.
@@ -17,7 +17,7 @@ export class EventDispatcher {
 	 * @memberof EventDispatcher
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
-	public readonly handlers: Array<EventHandler<any>>;
+	public handlers: Array<EventHandler<any>>;
 
 	/**
 	 * Event name.
@@ -57,14 +57,14 @@ export class EventDispatcher {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public async dispatch<Event extends EventPayload>(
-		event: Event
+		event: Event,
 	): Promise<EventDispatcherResponse> {
 		if (event.name !== this.name) {
-			return undefined;
+			return [];
 		}
 
 		if (this.handlers.length === 0) {
-			return undefined;
+			return [];
 		}
 
 		return Promise.allSettled(
@@ -74,7 +74,7 @@ export class EventDispatcher {
 				} catch (err: any) {
 					return Promise.reject(err);
 				}
-			})
+			}),
 		);
 	}
 
@@ -89,7 +89,7 @@ export class EventDispatcher {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public register<Event extends EventPayload>(
-		handler: EventHandler<Event>
+		handler: EventHandler<Event>,
 	): boolean {
 		if (this.handlers.find(h => h === handler)) {
 			return false;
@@ -110,9 +110,8 @@ export class EventDispatcher {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public unregister<Event extends EventPayload>(
-		handler: EventHandler<Event>
+		handler: EventHandler<Event>,
 	): boolean {
-		// const index = this.handlers.indexOf(handler);
 		const index = this.handlers.findIndex(h => h === handler);
 
 		if (index === -1) {
@@ -137,7 +136,7 @@ export class EventDispatcher {
 			return false;
 		}
 
-		this.handlers.splice(0, this.handlers.length);
+		this.handlers = [];
 		return true;
 	}
 }